requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==4.0.3
attrs==25.1.0
blinker==1.8.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
frozenlist==1.5.0
greenlet==3.1.1
groq==0.16.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.17
langchain-core==0.2.43
langchain-groq==0.1.10
langchain-text-splitters==0.2.4
langsmith==0.1.147
MarkupSafe==2.1.5
multidict==6.1.0
numpy==1.24.4
orjson==3.10.15
packaging==24.2
propcache==0.2.0
pydantic==2.10.6
pydantic-core==2.27.2
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.37
tenacity==8.5.0
typing-extensions==4.12.2
urllib3==2.2.3
werkzeug==3.0.6
yarl==1.15.2
zipp==3.20.2
